requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# LLM and Embeddings
openai==1.3.7
langchain==0.0.340
langchain-community==0.0.1
sentence-transformers==2.2.2

# Vector Databases
pinecone-client==2.2.4
faiss-cpu==1.7.4
chromadb==0.4.17

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
sqlite3

# Messaging APIs
requests==2.31.0
flask==3.0.0
slack-sdk==3.26.1
twilio==8.10.0

# Utilities
pandas==2.1.3
numpy==1.25.2
tiktoken==0.5.1
